requirements.txt
google-generativeai>=0.8.6
streamlit>=1.28.0
